+++ lib/src/core/library.dart
@@ -6,11 +6,11 @@ export "dimension.dart";
 
 export "precision/precision.dart";
 
+export "../derived/derived_templates.dart";
+
 part "extension_num.dart";
 part "unit.dart";
-part "unit2.dart";
 part "unit2_builder.dart";
-part "unit3.dart";
 part "measurement_prefix.dart";
 part "measurement.dart";
 part "measurement2.dart";
@@ -18,3 +18,4 @@ part "measurement3.dart";
 part "measurement2_builder.dart";
 
 part "../measurements/volume.dart";
+part "../generated/derived_units.dart";