+++ CHANGELOG.md
@@ -7,7 +7,7 @@ types you need to work with from the rest of your code.
 
 ### Added
 
-- Code generator support for derived units
+- Code generator support for derived units of any dimension, including custom units
 
 ### Changed